pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atlas-creator"
version = "0.1.0"
description = "Локальный инструмент для сборки spritesheet из видео с удалением фона через rembg"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "PySide6>=6.8",
  "Pillow>=11.0.0",
  "rembg[cpu]>=2.0.60",
  "numpy>=1.26.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]

[project.scripts]
atlas-creator = "src.main:run"

[tool.setuptools]
packages = ["src", "src.services", "src.ui"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
